Gtk: Update Button so it trims text to fit button with ellipsize.
Fixes #1722
